requirements.txt
pytz
pandas
requests
numpy
plotly
ta
typing-extensions
telebot
fastapi
apscheduler
uvicorn
kaleido
pymongo
certifi
python-multipart
jinja2
dash
dash-bootstrap-components
python-binance